requirements.txt
python-dotenv
pandas
pydantic